feat: Introducing activity feed (#219)
The Activity Feed is designed to give you a comprehensive and immediate overview of all the actions that have taken place within your organization. Whether it's an update to a schema, the addition of a note or memo, or any other significant event, the Activity Feed will have it recorded.

This is a draft feature, a work-in-progress MVP. We'd love to hear your feedback as we continue to make improvements.